% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_statedist}
\alias{calc_statedist}
\title{State distribution}
\usage{
calc_statedist(delta, Gamma)
}
\arguments{
\item{delta}{initial distribution of the Markov chain - vector of length \code{n_states}}

\item{Gamma}{transition matrices of the Markov chain - array of dimension \code{c(n_states, n_states, nObs)}}
}
\value{
all unconditional state distributions of the Markov chain
}
\description{
Computation of the unconditional state distribution (written in C++)
}
\examples{
delta <- c(0.5, 0.5)
Gamma <- array(c(0.9, 0.1, 0.1, 0.9), dim = c(2, 2, 10))
calc_statedist(delta, Gamma)
}
