% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hmm.R
\name{getPalette}
\alias{getPalette}
\title{Discrete colour palette for states}
\usage{
getPalette(n_states)
}
\arguments{
\item{n_states}{Number of states}
}
\value{
Vector of colours, of length n_states
}
\description{
Discrete colour palette for states
}
