% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hmm.R
\name{process_hmm}
\alias{process_hmm}
\title{Processes HMM model object}
\usage{
process_hmm(mod)
}
\arguments{
\item{mod}{model object as returned by \code{moveHMM}, \code{momentuHMM} or \code{hmmTMB}}
}
\value{
list containing original covariates \code{covs} and a function \code{predict_tpm}
to predict the transition probability matrix for new covariate values
}
\description{
Extracts covariates and sets up a prediction function for the transition probability matrix
}
\examples{
# no examples yet
}
