% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statedist.R
\name{statedist}
\alias{statedist}
\title{State distribution}
\usage{
statedist(mod, method = "gamlss")
}
\arguments{
\item{mod}{model object as returned by \code{moveHMM}, \code{momentuHMM} or \code{hmmTMB}}

\item{method}{method to use for computing the state distribution.

Options are "gamlss", "ar", "gan" and "bootstrap". See `Details'.}
}
\value{
statedist object containing the results and a plotting function
}
\description{
Computes the covariate-dependent state distribution for all covariates based on fitted HMM model.
}
\details{
Additional details...
}
\examples{
# no examples yet
}
